As a general scenario, we consider a Lindblad master equations denoted by,

$\mathcal{L}\rho =\frac{d \rho}{dt} = -i[H, \rho] + \sum_{k=1}^r L_k \rho L_k^\dagger -\frac{1}{2}\left\{L^\dagger_k L_k, \rho \right\}.$

From the above equation we introduce $p\leq r$ counting fields with weights $\nu_k$. This defines the generalized master equation (GME),

$\mathcal{L}_\chi \rho_\chi = \left(\mathcal{L} + \Delta \mathcal{L}_\chi \right)\rho_\chi,$

where,

$\Delta \mathcal{L}_\chi = \sum_{k=1}^p(1-e^{i\nu_k \chi}) L_k \rho L_k^\dagger .$

Whatever method we use to compute the FCS, we start by vectorizating the GME,

$\mathcal{L}_\chi\rho_\chi \to \mathcal{L}_\chi|\rho_\chi \rangle \rangle$

QuantumFCS.fcscumulants_recursive — Function

fcscumulants_recursive(H, J, mJ, nCum; <keyword arguments>)

Calculate n-th zero-frequency cumulant of full counting statistics using a recursive scheme.

Arguments

H: Arbitrary operator specifying the Hamiltonian.

J: Vector containing all jump operators which can be of any arbitrary operator type.

mJ: Vector containing the monitored jumps

tol=1e-3: Tracedistance used as termination criterion.

hmin=1e-7: Minimal time step used in the time evolution.

rates=ones(N): Vector or matrix specifying the coefficients for the jump operators.

Jdagger=dagger.(Jdagger): Vector containing the hermitian conjugates of the jump operators. If they are not given they are calculated automatically.

nu: vector of length 2*length(J) weights for each jump operator. By default down jumps, J, have weight -1 and up-jumps have weight +1.

kwargs...: Further arguments are passed on to the ode solver.