Documentation

Introduction

As a general scenario, we consider a Lindblad master equations,

$\mathcal{L}\rho =\frac{d \rho}{dt} = -i[H, \rho] + \sum_{k=1}^r L_k \rho L_k^\dagger -\frac{1}{2}\left\{L^\dagger_k L_k, \rho \right\}.$

We introduce $p \leq r$ counting fields $N_k$ with weights $\nu_k$. This lets us define the total current,

$N(t) = \sum_k \nu_k N_k(t) ~ .$

We further define the n-resolved density matrix $\rho_n(t)$ whose trace equals the probability to have accumulated $n$ jumps at time $t$ , $P(n,t) = \text{Tr}\left[ \rho_n(t) \right]$. Summing over the set of allowed values $\mathcal{N}$ for the total charge $N$ , we retrieve the standard density matrix,

$\rho = \sum_{n \in \mathcal{N}} \rho_n(t)~.$

We now consider the Fourier transform of the n-resolved density matrix,

$\rho_{\chi}(t) = \sum_{n \in \mathcal{N}} e^{i n \chi} \rho_n(t)~.$

$\chi$ is called the counting field and the timeevolution of $\rho_{\chi}(t)$ is given by the generalized master equation (GME),

$\mathcal{L}_\chi \rho_\chi = \left(\mathcal{L} + \delta \mathcal{L}_\chi \right)\rho_\chi,$

where,

$\delta \mathcal{L}_\chi = \sum_{k=1}^p(1-e^{i\nu_k \chi}) L_k \rho L_k^\dagger .$

Computing Cumulants using recursive Methods

We are ultimately interested in the $n$-th cumulant $\langle \langle I^n \rangle \rangle$ of the stochastic current,

$I(t) = \frac{dN}{dt}~,$

which we compute through the following recursive scheme,

$\langle \langle I^n \rangle \rangle = \sum_{m=1}^n \binom{n}{m} \langle \langle \mathbb{1} | \mathcal{L}^{(m)} | \rho_{\text{ss}}^{(n-m)}(\chi) \rangle \rangle ~,$

with the constituents,

$| \rho_{\text{ss}}^{(n)}(\chi) \rangle \rangle = \mathcal{L}^+ \sum_{m=1}^n \binom{n}{m} \left( \langle \langle I^m \rangle \rangle - \mathcal{L}^{(m)}\right) | \rho_{\text{ss}}^{n-m} \rangle \rangle ~,$

$\mathcal{L}^{(n)} = \left. \left( -i \partial_{\chi} \right)^n \mathcal{L}_\chi \right|_{\chi \to 0}~,$

and $\mathcal{L}^+$ being the Drazin inverse of $\mathcal{L}$.

Functions

QuantumFCS.fcscumulants_recursive — Function

fcscumulants_recursive(H, J, mJ, nC; <keyword arguments>)

Calculate n-th zero-frequency cumulant of full counting statistics using a recursive scheme.

Arguments

H: Arbitrary operator specifying the Hamiltonian.

J: Vector containing all jump operators which can be of any arbitrary operator type.

mJ: Vector containing the monitored jump operators.

Jdagger=dagger.(Jdagger): Vector containing the hermitian conjugates of the jump operators. If they are not given they are calculated automatically.

nu: vector of length 2*length(J) weights for each jump operator. By default down jumps, J, have weight -1 and up-jumps have weight +1.

kwargs...: Further arguments are passed on to the ode solver.

QuantumFCS.drazin — Function

drazin(H, J; rho_ss = steadystate.eigenvector(H, J))

Calculate the Drazin inverse of a Liouvillian defined by the Hamiltonian H and jump operators J.

Arguments

H: Arbitrary operator specifying the Hamiltonian.

J: Vector containing all jump operators which can be of any arbitrary operator type.

rho_ss: Density matrix specifying the steady-state of the Liouvillian. By default, it is found through steadystate.eigenvector. For large matrices the steady-state should be provided, as the best steady-state solver could vary.

QuantumFCS.m_jumps — Function

m_jumps(mJ; n=1; nu = vcat(fill(-1, Int(length(J)/2)),fill(1, Int(length(J)/2))))

Calculate the vectorized super-operator ℒ(n) = ∑ₖ (νₖ)ⁿ (Lₖ*)⊗Lₖ.

Arguments

mJ: List of monitored jumps

n : Power of the weights νₖ. By default set to 1, since this case appears more often.

nu: Set of weights νₖ, by default set to -1 for emission and +1 for absorption jumps.

QuantumFCS.drazin_apply — Function

drazin_apply()

Calculates the vector resulting from the Drazin inverse being applied to another vector.

Arguments

H: Arbitrary operator specifying the Hamiltonian.

J: Vector containing all jump operators which can be of any arbitrary operator type.

ρss : steady state of the system as a density matrix

α : input state as a density matrix

Examples